pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "document-adapter"
version = "0.4.0"
description = "LLM-friendly document template editing (DOCX/PPTX/HWPX) with MCP server and Claude API tool-use support"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Son Seongjun", email = "sonsj97@plateer.com" },
]
keywords = [
    "mcp",
    "model-context-protocol",
    "llm",
    "claude",
    "anthropic",
    "tool-use",
    "agent",
    "document",
    "docx",
    "pptx",
    "hwpx",
    "hwp",
    "template",
    "template-engine",
    "office",
    "office-automation",
    "word",
    "powerpoint",
    "hancom",
    "korean",
    "docxtpl",
    "python-docx",
    "python-pptx",
    "lxml",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Office/Business",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: Markup",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Environment :: Console",
]

dependencies = [
    "python-docx>=1.2",
    "docxtpl>=0.20",
    "python-pptx>=1.0",
    "lxml>=5.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40"]
# python-hwpx는 Non-Commercial License이므로 runtime 의존성에서 제외.
# 테스트에서만 fixture 생성용으로 사용 (자체 HwpxAdapter는 hwpx_core 기반이라 불필요).
dev = ["pytest>=8", "python-hwpx>=2.9"]

[project.scripts]
document-adapter-mcp = "document_adapter.mcp_server:main_sync"

[project.urls]
Homepage = "https://github.com/PlateerLab/document-adapter"
Repository = "https://github.com/PlateerLab/document-adapter"
Issues = "https://github.com/PlateerLab/document-adapter/issues"
PyPI = "https://pypi.org/project/document-adapter/"
Changelog = "https://github.com/PlateerLab/document-adapter/releases"

[tool.setuptools.packages.find]
include = ["document_adapter*"]
exclude = ["tests*", "examples*"]
